+++ docs/news.rst
@@ -17,6 +17,7 @@ Psycopg 3.1.11 (unreleased)
   usage (:ticket:`#628`).
 - Fix integer overflow in C/binary extension with OID > 2^31 (:ticket:`#630`).
 - Fix loading of intervals with days and months or years (:ticket:`#643`).
+- Work around excessive CPU usage on Windows (reported in :ticket:`#645`).
 - Fix building on Solaris and derivatives (:ticket:`#632`).
 - Fix possible lack of critical section guard in async
   `~AsyncCursor.executemany()`.